requirements.txt
Django==5.2.3
python-dotenv==1.0.0
requests==2.32.4
ldap3
Pillow
